requirements.txt

fastapi>=0.95.0
uvicorn[standard]>=0.21.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.2
requests>=2.28.0
jinja2>=3.0
python-multipart
starlette>=0.26.0
